stmt-sink: CFaderLight::ApplyInit 92.25 -> 96.12, and a 16-bit ~enum mask bug
CFaderLight::ApplyInit 0x1804a0 92.25 -> 96.12. Third instance of the same
contextual sink: pair 1 written as an alias temp made cl sink pair 2 as well;
writing both as if/else gives retail's (sunk, duplicated) block pair, 29/29 blocks.

`m_stateFlags &= ~SPRITE_STATE_HIDDEN` assembled to `and dword ptr [..],0xfffe`,
not `0xfffffffe` - MSVC 5.0 promotes the anonymous enum GZ_ENUM_FLAGS_BEGIN emits
to a 16-bit type, so the complement clears bits 16..31 as well. It is a real
CORRECTNESS bug, not only three bytes of encoding: every other flag in that word
is destroyed on every hide-cursor path. Fixed the three sites in Play.cpp with
`~static_cast<i32>(...)` (the spelling DinMgr2 already uses); retail's `83 60 40
fe` now matches byte for byte. The same expression appears at ~45 more sites
across 14 files owned by other lanes - see the report.

+++ src/Gruntz/Play.cpp
@@ -3742,7 +3742,7 @@ i32 CPlay::LoadCursorSprites(i32 frame, i32 flag) {
             return 0;
         }
         if (this->m_scrollSink != NULL) {
-            this->m_scrollSink->m_stateFlags &= ~SPRITE_STATE_HIDDEN;
+            this->m_scrollSink->m_stateFlags &= ~static_cast<i32>(SPRITE_STATE_HIDDEN);
         }
         this->m_cursorOffset.m_x = 0x10;
         this->m_cursorOffset.m_y = 0x10;
@@ -4172,7 +4172,7 @@ i32 CPlay::HandleDragMove(i32 a, i32 x, i32 y) {
                 }
             } else {
                 if (m_scrollSink != NULL) {
-                    m_scrollSink->m_stateFlags &= ~SPRITE_STATE_HIDDEN;
+                    m_scrollSink->m_stateFlags &= ~static_cast<i32>(SPRITE_STATE_HIDDEN);
                 }
             }
             CGameLevel* v = m_world->m_level;
@@ -4628,7 +4628,7 @@ i32 CPlay::ResetPlayState() {
         }
     }
     if (m_scrollSink != NULL) {
-        m_scrollSink->m_stateFlags &= ~SPRITE_STATE_HIDDEN;
+        m_scrollSink->m_stateFlags &= ~static_cast<i32>(SPRITE_STATE_HIDDEN);
     }
     m_inGame = 0;
     if (!PlaceStartGruntz()) {
